(trunk) Filled polygons now use stencil buffer to generate polygon interior fill mask - reduces number of render target switches required compared to using alpha blending to generate fill mask.
git-svn-id: http://svn.gplates.org/projects/gplates/trunk@13952 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/gui/Globe.cc
@@ -444,7 +444,7 @@ GPlatesGui::Globe::render_globe_hemisphere_surface(
 	renderer.gl_clear_stencil(); // Clear stencil to 0
 	// NOTE: Depth/stencil writes must be enabled for depth/stencil clears to work.
 	renderer.gl_depth_mask(GL_TRUE);
-	renderer.gl_stencil_mask(~GLuint(0)/*all ones*/);
+	renderer.gl_stencil_mask(~0/*all ones*/);
 	renderer.gl_clear(GL_DEPTH_BUFFER_BIT | GL_STENCIL_BUFFER_BIT);
 
 	// Enable depth testing but disable depth writes (for the grid lines).
@@ -538,7 +538,7 @@ GPlatesGui::Globe::render_globe_sub_surface(
 	renderer.gl_clear_stencil(); // Clear stencil to 0
 	// NOTE: Depth/stencil writes must be enabled for depth/stencil clears to work.
 	renderer.gl_depth_mask(GL_TRUE);
-	renderer.gl_stencil_mask(~GLuint(0)/*all ones*/);
+	renderer.gl_stencil_mask(~0/*all ones*/);
 	renderer.gl_clear(GL_DEPTH_BUFFER_BIT | GL_STENCIL_BUFFER_BIT);
 
 	// Draw the sub-surface geometries.
